Move trace objects to traces module and rename `y_shifted` to `y_base`
